% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{achievement_levels}
\alias{achievement_levels}
\title{CAAQS Achievement levels for air pollutaionts}
\format{A data frame with 8 variables:
\describe{
  \item{parameter}{The name of the parameter}
  \item{labels}{Achieved or Not Achieved}
  \item{lower_breaks}{Lower concentration in the category}
  \item{upper_breaks}{Upper concentration in the category}
  \item{units_html}{The units in html codes}
  \item{units_unicode}{The units in unicode}
  \item{val_labels_html}{Labels for the category values in html}
  \item{val_labels_unicode}{Labels for the category values in unicode}
  \item{colour}{The colour of the category (in hexadecimal code)}
  
}}
\usage{
achievement_levels
}
\description{
CAAQS Achievement levels for air pollutaionts
}
\keyword{datasets}

