% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dates.R
\name{add_ymd}
\alias{add_ymd}
\title{Add a year, month, and day column to a dataframe based on a date column}
\usage{
add_ymd(df, datecol, tz = "Etc/GMT+8", outnames = NULL)
}
\arguments{
\item{df}{dataframe}

\item{datecol}{the column in df that contains the dates as a character string}

\item{tz}{the timezone}

\item{outnames}{the names for the new columns in order of year, month, day
(defaults to "year", "month", "day")}
}
\value{
a dataframe with the new columns added
}
\description{
Add a year, month, and day column to a dataframe based on a date column
}

