% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airzone_map}
\alias{airzone_map}
\title{Borders of British Columbia Airzones}
\format{A SpatialPolgonsDataFrame with 7 polygons and 1 variable
\describe{
\item{Airzone}{The name of the airzone}
}}
\usage{
airzone_map
}
\description{
Borders of British Columbia Airzones
}
\keyword{datasets}

