% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airzone_metric.R
\name{airzone_metric}
\alias{airzone_metric}
\title{Calculate the airzone metric and achievement status}
\usage{
airzone_metric(df, n, avg, stn_id, stn_name, std)
}
\arguments{
\item{df}{The dataframe}

\item{n}{The column containing the number of years each 3yr avg is based on}

\item{avg}{The column containing the 3 yr averages}

\item{stn_id}{the column containing the station ids}

\item{stn_name}{the column containing the station names}

\item{std}{the standard (e.g. 63 for ozone)}
}
\value{
A dataframe with two columns: metric value and achievement status
}
\description{
<full description>
}

