% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airzone_metric.R
\name{assign_airzone}
\alias{assign_airzone}
\title{Assign locations to airzones}
\usage{
assign_airzone(data, airzones, az = "Airzone", station_id = "ems_id",
  coords = c("lat", "lon"))
}
\arguments{
\item{data}{Data frame. Contains station ids and air quality data}

\item{airzones}{sf PPOLYGON or MULTIPOLYGON reflecting airzone
locations}

\item{az}{Character. Name of airzones column in the 'airzones'
sf object}

\item{station_id}{Character. Name of the station_id column in 'data'}

\item{coords}{Character vector. Names of the columns containing longitude and
latitude (respectively) for each station. Defaults to 'lon' and 'lat'}
}
\description{
With a data set containing station locations, calculate which airzones each
station belongs to. Requires an sf object from the
\code{sf} package containing the airzone names and locations (polygons).
}
\details{
For British Columbia, consider using the \code{airzones()} function to
get an sf object of BC airzones from the \code{\link[bcmaps]{bcmaps}} package.
}
\examples{

\dontrun{
# Using the airzones function from the bcmaps package
bc_airzones <- bcmaps::airzones()

by_airzones <- assign_airzone(data, bc_airzones)
}

}
