% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{caaqs_extractors}
\alias{caaqs_extractors}
\alias{extract_caaqs}
\alias{extract_daily}
\alias{extract_yearly}
\alias{extract_three_yr_rolling}
\title{Extract results from a caaqs analysis}
\usage{
extract_caaqs(x)

extract_daily(x)

extract_yearly(x)

extract_three_yr_rolling(x)
}
\arguments{
\item{x}{An object of class \code{"caaqs"} (result of running a \code{*_caaqs()} function)}
}
\value{
Varies according to the function and the parameter being analyzed:

\code{extract_caaqs()}:
\itemize{
\item caaqs results
}

\code{extract_daily()}:
\itemize{
\item pm2.5_24h, pm2.5_annual: Daily average
\item o3, so2_3yr, no2_3yr: Daily maximum
\item so2_1yr, no2_1yr: None
}

\code{extract_yearly()}:
\itemize{
\item pm2.5_24h: Annual 98th percentile daily average
\item pm2.5_annual: Annual average of daily averages
\item o3: Annual 4th highest daily maximum
\item so2_1yr, no2_1yr: Annual average of hourly values
\item so2_3yr: Annual 99th percentile or daily maximums
\item no2_3yr: Annual 98th percentile or daily maximums
}

\code{extract_three_yr_rolling()}
\itemize{
\item pm2.5_24h, pm2.5_annual, o3, so2_3yr, no2_3yr: Three-year rolling average of the
yearly value.
\item so2_1yr, no2_1yr: None
}
}
\description{
Extract results from a caaqs analysis
}
