% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caaqs.R
\name{caaqs_management}
\alias{caaqs_management}
\title{Calculate CAAQS management levels excluding days with Exceptional Events
or Transboundary Flows}
\usage{
caaqs_management(x, exclude_df = NULL, exclude_df_dt = NULL,
  quiet = FALSE)
}
\arguments{
\item{x}{an object of class \code{caaqs}}

\item{exclude_df}{Data frame. The dates over which data should be excluded
(see details). Data should be arranged either with one column of dates to
omit, or two columns specifying a series of start and end date ranges to
omit.}

\item{exclude_df_dt}{Character vector. The names of the date columns in
\code{exclude_df}. Must specify either one (a series of dates), or two (the
start and end columns specifying dates ranges).}

\item{quiet}{Logical. Suppress progress messages (default FALSE)}
}
\value{
object of class \code{caaqs}
}
\description{
Calculate CAAQS management levels excluding days with Exceptional Events
or Transboundary Flows
}
\details{
To omit days which are suspected to be influenced by Transboundary
Flows or Exceptional Events create a data frame that either a) contains a
column listing all the days which are to be omitted, or b) contains two
columns listing the start and end dates of all the date periods which are
to be omitted. This is supplied as \code{exclude_df}. Use
\code{exlcude_df_dt} to specify the name of the column containing the
dates, or the names of the columns containing the start and end of the date
ranges (see examples and vignette for more details).
}
