% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caaqs.R
\name{caaqs_metric}
\alias{caaqs_metric}
\alias{pm_24h_caaq}
\alias{pm_annual_caaq}
\alias{o3_caaq}
\alias{so2_1yr_caaq}
\alias{so2_3yr_caaq}
\alias{no2_1yr_caaq}
\alias{no2_3yr_caaq}
\title{Compute specific CAAQ metrics}
\usage{
pm_24h_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)

pm_annual_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)

o3_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)

so2_1yr_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)

so2_3yr_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)

no2_1yr_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)

no2_3yr_caaq(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL, quiet = FALSE)
}
\arguments{
\item{data}{Data frame. Hourly raw pollution data with at least date and
value columns}

\item{dt}{Character. The name of the date-time column. Default \code{"date"}}

\item{val}{Character. The name of the value column. Default \code{"value"}}

\item{by}{Character vector. Grouping variables in data, probably an id if 
using multiple sites. Even if not using multiple sites, you shoud specify 
the id column so that it is retained in the output.}

\item{exclude_df}{Data frame. The dates over which data should be excluded 
(see details). Data should be arranged either with one column of dates to 
omit, or two columns specifying a series of start and end date ranges to 
omit.}

\item{exclude_df_dt}{Character vector. The names of the date columns in 
\code{exclude_df}. Must specify either one (a series of dates), or two (the
start and end columns specifying dates ranges).}

\item{quite}{Logical. Suppress progress messages (default FALSE)}
}
\value{
Data frame with CAAQ metrics
}
\description{
Compte specific CAAQ metrics for different forms of air pollution.
}
\details{
To omit days which are suspected to be influenced by Transboundary
  FLows or Exceptional Events create a data frame that either a) contains a
  column listing all the days which are to be omitted, or b) contains two
  columns listing the start and end dates of all the date periods which are
  to be omitted. This is suppled as \code{exclude_df}. Use
  \code{exlcude_df_dt} to specify the name of the column containing the
  dates, or the names of the columns containing the start and end of the date
  ranges.
}
\examples{

\dontrun{
# Normal run
pm <- pm_24h_caaq(pm25_sample_data, by = c("ems_id", "site"))

# Exclude dates
high_dates <- data.frame(ems_id = "0310162",
                         site = "Port Moody Rocky Point Park", 
                         date = seq(as.Date("2012-06-11"),
                                    as.Date("2012-06-30"), by = "1 day"))
                                    
pm_ex <- pm_24h_caaq(pm25_sample_data, 
                     by = c("ems_id", "site"),
                     exclude_df = high_dates,
                     exclude_df_dt = "date")
}

}
\references{
CCME Guidance document on achievement determination Canadian ambient air quality standards for fine particulate matter and ozone \url{https://www.ccme.ca/files/Resources/air/aqms/pn_1483_gdad_eng.pdf}.
}
