% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_neg}
\alias{clean_neg}
\title{Replaces negative values in a vector of air quality readings with CAAQS-compliant values}
\usage{
clean_neg(x, type = c("pm25", "ozone", "so2", "nox"))
}
\arguments{
\item{x}{vector of air quality readings}

\item{type}{type of air quality readings; one of: "pm25", "ozone", "so2", "nox"}
}
\value{
a vector of the same length as x with negative values replaced with CAAQS compliant values
}
\description{
Replaces negative values in a vector of air quality readings with CAAQS-compliant values
}
\details{
For gases (type = "ozone", "so2", or "nox"),
         replaces values \eqn{\le} -5 with \code{NA} and values between -5 and 0 with 0.
         For type = "pm25", replaces values \eqn{\le} -3.5 with \code{NA} and values between -3.5 and 0 with 0.
}
\examples{
\dontrun{

}
}

