% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cut_caaq}
\alias{cut_caaq}
\title{Bin a vector of air quaility readings into CAAQS management levels}
\usage{
cut_caaq(x, metric, drop_na = FALSE)
}
\arguments{
\item{x}{vector of air quality readings (pollutant concentrations)}

\item{metric}{What is the metric? Must be one of: \code{"o3", "pm2.5_annual", "pm2.5_daily"}}

\item{drop_na}{Should NA values be dropped, or retained as factor level "No Data" (default)?}
}
\value{
factor
}
\description{
Can be used for Ozone, PM2.5 (daily or annual)
}
\examples{
\dontrun{

}
}

