% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut.R
\name{cut_management}
\alias{cut_management}
\title{Bin a vector of air quaility readings into CAAQS management levels}
\usage{
cut_management(x, parameter, output = "labels", drop_na = FALSE)
}
\arguments{
\item{x}{vector of air quality readings (pollutant concentrations)}

\item{parameter}{What is the parameter? Must be one of: \code{"o3",
"pm2.5_annual", "pm2.5_24h"}}

\item{output}{should the function output labels (\code{"labels"}; default or
break values in unicode (\code{"breaks_u"}) or break values in html
(\code{"breaks_h"}), or colour values (\code{"colour"} or \code{"color"})?}

\item{drop_na}{Should NA values be dropped, or retained as factor level
"Insufficient Data" (default)?}
}
\value{
factor
}
\description{
Can be used for Ozone, PM2.5 (24h or annual)
}

