% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{daily_max}
\alias{daily_max}
\title{Find the daily maximum of a value by some factor}
\usage{
daily_max(input, dat = "date", val = "value", by = c("ems_id", "site"),
  thresh = 63, n.readings.min = 18)
}
\arguments{
\item{input}{Dataframe}

\item{dat}{character the column containing dates}

\item{val}{character the value column}

\item{by}{character the columns to group by}

\item{thresh}{numeric.}

\item{n.readings.min}{numeric.}
}
\description{
Given a dataframe with one value column, find the maximum value by some date.
}

