% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_stat.R
\name{daily_stat}
\alias{daily_stat}
\title{Find the daily maximum of a value by some factor}
\usage{
daily_stat(data, dt = "date", val = "value", by = c("ems_id", "site"),
  digits = 1, thresh, n_readings_min, stat)
}
\arguments{
\item{data}{Dataframe}

\item{dt}{character the column containing dates}

\item{val}{character the value column}

\item{by}{character the columns to group by}

\item{digits}{numeric.}

\item{thresh}{numeric.}

\item{n_readings_min}{numeric.}

\item{stat}{numeric.}
}
\description{
Given a dataframe with one value column, find the maximum value by some date.
}

