% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_stat.R
\name{daily_stat_page}
\alias{daily_stat_page}
\alias{o3_daily_max}
\alias{no2_daily_max}
\alias{so2_daily_max}
\alias{pm_daily_avg}
\title{Cacluate daily maxima or averages for different sensor readings.}
\usage{
o3_daily_max(data, dt = "date_time", val = "rolling8", by = NULL)

no2_daily_max(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

so2_daily_max(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

pm_daily_avg(data, dt = "date_time", val = "value", by = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column. Default
\code{"date_time"}}

\item{val}{the name (as a character string) of the PM2.5 value column. Default
\code{"value"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specfify
the id column so that it is retained in the output.}

\item{exclude_df}{a data.frame specifying which date ranges and sites to
exclude. The data.frame must have to one date column if specific dates or
date-times are to be excluded, and exactly two date or date-time columns if
a date range is needed. must have all the columns of by provided}

\item{exclude_df_dt}{specifies the one (or two) date columns to be use in the
exclude_df data.frame.}
}
\value{
data frame with the daily averages, can be input into
}
\description{
Cacluate daily maxima or averages for different sensor readings.
}
