% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{days_in_quarter}
\alias{days_in_quarter}
\title{Calcuate number of days in quarter}
\usage{
days_in_quarter(quarter, year)
}
\arguments{
\item{quarter}{the quarter of the year (1-4)}

\item{year}{the year as a numeric}
}
\value{
vector with the number of days in that quarter year combination.
}
\description{
Calcuate number of days in quarter
}
