% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{days_in_year}
\alias{days_in_year}
\title{Calcuate number of days in year}
\usage{
days_in_year(year)
}
\arguments{
\item{year}{the year as a numeric}
}
\value{
vector with the number of days in that year.
}
\description{
Calcuate number of days in year
}
