% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eetf.R
\name{eetf}
\alias{eetf}
\title{Flag Execeptional Events or Transboundary Flows affecting air quality}
\usage{
eetf(daily_df, id_col = "ems_id", date_col = "date", eetf)
}
\arguments{
\item{daily_df}{data frame of daily air quality values}

\item{id_col}{column in daily_df that contains the id of air quality monitors}

\item{date_col}{"date" column in daily_df}

\item{eetf}{a data frame with at least two columns: \code{id} and \code{date}
that define the monitor IDs and dates for which EEs or TFs have 
been identified. The \code{date} column should be of class \code{Date}, 
\code{POSIXt}, or character in the form \code{"YYYY-MM-DD"}.}
}
\value{
data frame: daily_df with an additional column of logical values indicating EE/TFs.
}
\description{
Flag Execeptional Events or Transboundary Flows affecting air quality
}
\examples{
daily_avg_file <- system.file("examples", "daily_avg.csv", package = "rcaaqs")
eetf_data_file <- system.file("examples", "eetf.csv", package = "rcaaqs")
daily_avg <- read.csv(daily_avg_file, stringsAsFactors = FALSE)
eetf_data <- read.csv(eetf_data_file, stringsAsFactors = FALSE)
eetf(daily_df = daily_avg, id_col = "ems_id", date_col = "dates", eetf = eetf_data)
}

