% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_data.R
\name{exclude_data}
\alias{exclude_data}
\title{Given exclude rows from input dataset based on supplied data.frame which specifies specific dates or date ranges.}
\usage{
exclude_data(data, dt, by, exclusion_df, exclusion_date_cols)
}
\arguments{
\item{data}{Dataframe}

\item{dt}{character the column containing date-times}

\item{by}{an optional character vector specifying the columns to join on.}

\item{exclusion_df}{the data.frame that has all the columns in the by
parameter, in addition exactly one or two date columns.}

\item{exclusion_date_cols}{a character vector with exactly one or two date columns.}
}
\value{
dataframe with the specified dates or date ranges excluded.
}
\description{
Given exclude rows from input dataset based on supplied data.frame which specifies specific dates or date ranges.
}
