% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{filled_rolling_mean}
\alias{filled_rolling_mean}
\title{Calculate a rolling mean, but fill in missing date values with NA.}
\usage{
filled_rolling_mean(dat, val, interval = 3600, width = 8,
  valid_thresh = 6, digits = 1)
}
\arguments{
\item{dat}{the input date or date-time value}

\item{val}{some numeric value to be interpolated with NAs.}

\item{interval}{some numeric step size. for date-times, in seconds.}

\item{width}{to calculate the mean, look back from current date 
to width*interval behind.}

\item{valid_thresh}{If there are less than width values available, 
then the mean will still be calculated if there are at least valid_thresh available. 
Set to equal width if this behaviour is not desired.}

\item{digits}{how many digits should the value be rounded to?}
}
\value{
vector with all input dates and values, interpolated with NAs.
}
\description{
Calculate a rolling mean, but fill in missing date values with NA.
}
