% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caaqs_methods.R
\name{get_caaqs}
\alias{get_caaqs}
\alias{get_hourly}
\alias{get_daily}
\alias{get_yearly}
\alias{get_three_yr_rolling}
\title{Extract results from a caaqs analysis}
\usage{
get_caaqs(x)

get_hourly(x)

get_daily(x)

get_yearly(x)

get_three_yr_rolling(x)
}
\arguments{
\item{x}{An object of class \code{"caaqs"} (result of running a \code{*_caaqs()} function)}
}
\value{
Varies according to the function and the parameter being analyzed:

\code{get_caaqs()}:
\itemize{
\item caaqs results
}

\code{get_hourly()}:
\itemize{
\item so2_1yr, no2_1yr: Hourly values used for annual average
\item pm2.5_24h, pm2.5_annual, o3, so2_3yr, no2_3yr: None
}

\code{get_daily()}:
\itemize{
\item pm2.5_24h, pm2.5_annual: Daily average
\item o3, so2_3yr, no2_3yr: Daily maximum
\item so2_1yr, no2_1yr: None
}

\code{get_yearly()}:
\itemize{
\item pm2.5_24h: Annual 98th percentile daily average
\item pm2.5_annual: Annual average of daily averages
\item o3: Annual 4th highest daily maximum
\item so2_1yr, no2_1yr: Annual average of hourly values
\item so2_3yr: Annual 99th percentile of daily maximums
\item no2_3yr: Annual 98th percentile of daily maximums
}

\code{get_three_yr_rolling()}
\itemize{
\item pm2.5_24h, pm2.5_annual, o3, so2_3yr, no2_3yr: Three-year rolling average of the
yearly value.
\item so2_1yr, no2_1yr: None
}
}
\description{
Extract results from a caaqs analysis
}
\examples{
pm <- pm_24h_caaqs(pm25_sample_data, by = c("ems_id", "site"))

get_caaqs(pm)

# Extract intermediate objects:

get_daily(pm)
get_yearly(pm)
get_three_yr_rolling(pm)

}
