% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{get_colours}
\alias{get_colours}
\title{Get a vector of colours for management levels or achievement levels}
\usage{
get_colours(type = "management", drop_na = FALSE)
}
\arguments{
\item{type}{"achievement" or "management"}

\item{drop_na}{Should NA (i.e., Insufficient Data) be excluded (\code{TRUE})
or includeded \code{FALSE} (the default)?}
}
\value{
named vector of colours (hex)
}
\description{
Get a vector of colours for management levels or achievement levels
}
