% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut.R
\name{get_labels}
\alias{get_labels}
\title{Get labels}
\usage{
get_labels(x, output, sub_na)
}
\arguments{
\item{x}{dataframe - a subset of achievement_levels or caaqs_levels}

\item{output}{one of: "labels", "breaks_h", "breaks_u", "colour", "color"}
}
\value{
character vector
}
\description{
Get labels
}
\keyword{internal}

