% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cut_caaq.R
\name{get_levels}
\alias{get_levels}
\title{Return CAAQS management levels for any, or all, parameters}
\usage{
get_levels(parameter = "all")
}
\arguments{
\item{parameter}{one of "all" (default), "o3", "PM2.5_annual", "PM2.5_daily"}
}
\value{
A data frame
}
\description{
Return CAAQS management levels for any, or all, parameters
}
\examples{
\dontrun{
get_levels("o3")
}
}

