% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.R
\name{get_opts}
\alias{get_opts}
\title{Get Options}
\usage{
get_opts(name = NULL)
}
\arguments{
\item{name}{option name (currently only the timezone, \code{"tz"}). If
  \code{NULL} (the default), will return a list of all options.}
}
\value{
a list of options specified (all if no arguments given)
}
\description{
Get Options
}
\examples{
get_opts()
get_opts("tz")
}

