% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{get_std}
\alias{get_std}
\title{get the CAAQS value for a given parameter}
\usage{
get_std(parameter = "all")
}
\arguments{
\item{parameter}{parameter of interest. One or more of}
}
\value{
a named vector of standards
}
\description{
get the CAAQS value for a given parameter
}
\keyword{internal}
