% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{max_na}
\alias{max_na}
\title{Find the maximum, return NA if no non-NA values.}
\usage{
max_na(x)
}
\arguments{
\item{x}{numeric}
}
\description{
max typically returns -Inf (and a warning) if no non-NA values; this wrapper avoids that.
}

