% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{mean_na}
\alias{mean_na}
\title{Find the mean, return NA if no non-NA values.}
\usage{
mean_na(x)
}
\arguments{
\item{x}{numeric}
}
\description{
mean typically returns a warning if no non-NA values; this wrapper avoids that.
}
