% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{mid_breaks}
\alias{mid_breaks}
\title{Move annual breaks to the midpoint of the year}
\usage{
mid_breaks(width = "1 year")
}
\arguments{
\item{width}{The desired interval of the breaks}
}
\value{
a function
}
\description{
Move annual breaks to the midpoint of the year
}

