% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{n.within.window}
\alias{n.within.window}
\title{number within a rolling window and step size (for dates or numeric)}
\usage{
n.within.window(x, interval, window)
}
\arguments{
\item{x}{the date, date_time, or numeric value}

\item{interval}{some numeric step size. for date-times, in seconds.}

\item{window}{some numeric look-back window.}
}
\value{
vector with all input dates and values, interpolated with NAs.
}
\description{
number within a rolling window and step size (for dates or numeric)
}

