% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3_ann_4th_highest.R
\name{o3_ann_4th_highest}
\alias{o3_ann_4th_highest}
\title{Calculate the 4th highest annual ozone reading, following CAAQS data completeness 
rules}
\usage{
o3_ann_4th_highest(data, dt = "date", val = "max8hr", by = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column. Default
\code{"date_time"}}

\item{val}{the name (as a character string) of the ozone value column. Default
\code{"value"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specify
the id column so that it is retained in the output.}
}
\value{
data frame with the fourth highest maximum daily reading, can be input into 
 \code{\link{o3_three_yr_avg}}
}
\description{
Calculate the 4th highest annual ozone reading, following CAAQS data completeness 
rules
}
\seealso{
\code{\link{o3_three_yr_avg}}
}

