% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caaqs.R
\name{o3_caaq}
\alias{o3_caaq}
\title{Calculate the Ozone CAAQ metric}
\usage{
o3_caaq(data, year = "year", val = "ozone_metric", by = NULL)
}
\arguments{
\item{data}{Data frame. Contains three year average data output from
\code{o3_three_yr_avg}}

\item{year}{Character. The name of the data column containing years. Defaults
to 'year'.}

\item{val}{Character. The name of the data column containing the ozone metric.
Defaults to 'ozone_metric'.}

\item{by}{Character. The name(s) of the data column(s) specifying the data
grouping variables (i.e. site_id, etc.). Even if not using multiple sites,
you shoud specfify the id column so that it is retained in the output.}
}
\value{
A data frame arranged by grouping variables (if present) with caaq
  metrics, achievement levels and management levels.
}
\description{
Calculates and returns the Ozone CAAQ metric based on a rolling three-year 
average.
}
\examples{

data(o3_sample_data)
o3_rolling <- o3_rolling_8hr_avg(o3_sample_data, by = c("ems_id", "site"))
o3_daily <- o3_daily_max(o3_rolling, by = c("ems_id", "site"))
o3_4th_highest <- o3_ann_4th_highest(o3_daily, by = c("ems_id", "site"))
o3_avg <- o3_three_yr_avg(o3_4th_highest, by = c("ems_id", "site"))

o3_final_caaq <- o3_caaq(o3_avg, by = c("ems_id", "site"))

}
