% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o3_sample_data}
\alias{o3_sample_data}
\title{Sample hourly ozone data for 3 years for 4 stations}
\format{
A data frame with 6 variables:
\describe{
\item{ems_id}{id of the monitoring station}
\item{date_time}{date/time (POSIXct) in ETC/GMT+8 timezone}
\item{site}{name of the monitoring station}
\item{year}{year}
\item{month}{month}
\item{parameter}{name of the parameter ("O3")}
\item{value}{ozone hourly value}
\item{units}{the units of value -- always 'ppb'}
}
}
\usage{
o3_sample_data
}
\description{
Sample hourly ozone data for 3 years for 4 stations
}
\keyword{datasets}
