% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3_three_yr_avg.R
\name{o3_three_yr_avg}
\alias{o3_three_yr_avg}
\title{Calculate the 3-year rolling average of the ozone reading, following CAAQS data completeness 
rules}
\usage{
o3_three_yr_avg(data, dt = "date", val = "max8hr", by = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column. Default
\code{"date_time"}}

\item{val}{the name (as a character string) of the ozone value column. Default
\code{"value"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specify
the id column so that it is retained in the output.}
}
\value{
data frame with the 3-year rolling averages
}
\description{
Calculate the 3-year rolling average of the ozone reading, following CAAQS data completeness 
rules
}

