% Generated by roxygen2 (4.0.2): do not edit by hand
\name{output_geojson}
\alias{output_geojson}
\title{Output a spatial*DataFrame to geojson}
\usage{
output_geojson(spdf, outpath, outname)
}
\arguments{
\item{spdf}{The SpatialDataFrame}

\item{outpath}{The output path}

\item{outname}{The output filename}
}
\value{
Nothing
}
\description{
Deals with a peculiarity of writeOGR where file has to be \code{".geojson"}, optionally preceeded by a path, but cannot have a filename preceeding the \code{".geojson"}.
Also deals with factors (including ordered factors)
}
\examples{
\dontrun{
output_geojson(myspatialdataframe, "out", "file.geojson")

}
}

