% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{pad.date.time}
\alias{pad.date.time}
\title{Calcuate number of days in quarter}
\usage{
pad.date.time(dat, val, interval)
}
\arguments{
\item{dat}{the date, date_time, or numeric value}

\item{val}{some value to be interpolated}

\item{interval}{some numeric step size. for date-times, in seconds.}
}
\value{
data.frame with all input dates and values, interpolated with NAs.
}
\description{
Calcuate number of days in quarter
}

