% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse_pm_monitors.R
\name{parse_pm_monitors}
\alias{parse_pm_monitors}
\title{Parse multiple columns to extract a simplified PM2.5 monitor type from them.}
\usage{
parse_pm_monitors(...)
}
\arguments{
\item{...}{Columns to parse, supplied in decreasing order of reliability}
}
\value{
a character vector with simplified PM2.5 monitor names
}
\description{
<full description>
}

