% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pm_data_complete.R
\name{percent_valid_days}
\alias{percent_valid_days}
\title{Given a vector of dates (in a single year), calculate the percentage of days in a quarter}
\usage{
percent_valid_days(dates, q = c("year", "Q1", "Q2", "Q3", "Q4"),
  tz = "Etc/GMT-8")
}
\arguments{
\item{dates}{a vector of dates}

\item{q}{the time period of interest, one of: "year","Q1","Q2","Q3","Q4"}

\item{tz}{the timezone the dates are in. Default Etc/GMT-8 with no Daylight savings}
}
\value{
A percentage of days in the specified quarter that are in the supplied vector
}
\description{
Given a vector of dates (in a single year), calculate the percentage of days in a quarter
}

