% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_caaqs}
\alias{plot_caaqs}
\title{Plot yearly CAAQS}
\usage{
plot_caaqs(
  x,
  id = NULL,
  id_col = NULL,
  year_min = NULL,
  year_max = NULL,
  plot_std = TRUE,
  plot_mgmt = TRUE,
  base_size = 10,
  annot_size = NULL
)
}
\arguments{
\item{x}{an object of class \code{caaqs_mgmt} (i.e. output of
\code{caaqs_management()})}

\item{id}{Character. Id of the station you would like to plot}

\item{id_col}{Character. Column in which to look for \code{id}}

\item{year_min}{Numeric. Minimum year to include. Data will be filtered or
filled to match.}

\item{year_max}{Numeric. Maximum year to include. Data will be filtered or
filled to match.}

\item{plot_std}{Logical. Should the CAAQs standard be plotted?}

\item{plot_mgmt}{Logical. Should the CAAQs management standards be plotted?}

\item{base_size}{Numeric. Base font size for the plot}

\item{annot_size}{Numierc. size of annotations. Scaling of this size is not
the same as the rest of the font sizes, so you will have to experiment.
Defaults to 0.32*base_size}
}
\value{
a ggplot2 object
}
\description{
Plot yearly CAAQS with exclusions for a particular station
}
