% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rolling}
\alias{plot_rolling}
\title{Plot rolling 3-yr averages}
\usage{
plot_rolling(
  x,
  id = NULL,
  id_col = NULL,
  plot_std = TRUE,
  plot_mgmt = TRUE,
  base_size = 10,
  annot_size = NULL
)
}
\arguments{
\item{x}{an object of class \code{caaqs_mgmt} (i.e. output of
\code{caaqs_management()})}

\item{id}{the id of the station you would like to plot}

\item{id_col}{the column in which to look for \code{id}}

\item{plot_std}{Should the CAAQs standard be plotted?}

\item{plot_mgmt}{Should the CAAQs management standards be plotted?}

\item{base_size}{base font size for the plot}

\item{annot_size}{size of annotations. Scaling of this size is not the same
as the rest of the font sizes, so you will have to experiment. Defaults to
0.32*base_size}
}
\value{
a ggplot2 object
}
\description{
Plot rolling 3-yr averages with exclusions for a particular station
}
