% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_station_instruments}
\alias{plot_station_instruments}
\title{Plot date ranges at which different instruments were deployed at each station}
\usage{
plot_station_instruments(data, dt = "date_time", station = "station_name",
  instrument = "instrument")
}
\arguments{
\item{data}{data frame}

\item{dt}{a date or date/time column in data}

\item{station}{column in data containing station names or ids}

\item{instrument}{column in data containing instrument type}
}
\value{
a ggplot2 opbject
}
\description{
Plot date ranges at which different instruments were deployed at each station
}
