% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pm_98_percentile.R
\name{pm_98_percentile}
\alias{pm_98_percentile}
\title{Calculate the annual 98th percentile of daily average PM2.5 values according
to CAAQS standards}
\usage{
pm_98_percentile(data, date = "date", val = "avg_24h", nr = "n_readings",
  by = NULL, std = 28, daily_valid = 18, type = "caaqs")
}
\arguments{
\item{data}{data frame}

\item{date}{the name of the "date" column (as a character string). Default
\code{"date"}}

\item{val}{the name of the column with daily average PM2.5 values. Default
\code{"avg_24h"}.}

\item{nr}{the column containing the number of readings per day (default
\code{"n_readings"}).}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specfify
the id column so that it is retained in the output.}

\item{std}{the value of the PM2.5 standard (default 28).}

\item{daily_valid}{The minimum number of hourly readings in a day for the
daily average to be considered valid (default \code{18}).}

\item{type}{type of 98th percentile calculation to use. Default "caaqs"}
}
\value{
A data frame with 98th percentiles of daily averages, per year
}
\description{
Also computes the data completeness criteria (optional). Designed to be used
with the output from \code{\link{pm_daily_avg}}.
}
\seealso{
\code{\link{pm_daily_avg}}, \code{\link{quantile2}}
}

