% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pm_annual_average.R
\name{pm_annual_average}
\alias{pm_annual_average}
\title{Calculate the annual average pm2.5 concentration,
optionally with completeness criteria}
\usage{
pm_annual_average(data, datecol, valcol, by = NULL, completeness = TRUE,
  year_valid = 75, q_valid = 60)
}
\arguments{
\item{data}{data frame}

\item{datecol}{the name of the "date" column (as a character string)}

\item{valcol}{the name of the column with daily average PM2.5 values}

\item{by}{character vector of grouping variables in data, probably an id if using multiple
sites. Even if not using multiple sites, you shoud specfify the id column
so that it is retained in the output.}

\item{completeness}{Should the completeness criteria be calculated (default
TRUE)}

\item{year_valid}{The percentage of valid days required in a year (default
75). Only required if calculating the completeness criteria.}

\item{q_valid}{The percentage of valid days required in each quarter
  (default 60). Only required if calculating the completeness criteria.}
}
\value{
a data frame with annual average values per year
}
\description{
Calculate the annual average pm2.5 concentration,
optionally with completeness criteria
}
\seealso{
\code{\link{pm_daily_avg}}
}

