% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pm_avg_daily}
\alias{pm_avg_daily}
\title{Calcualte the daily average PM2.5 reading, following CAAQS data completeness rules}
\usage{
pm_avg_daily(data, datecol, valcol)
}
\arguments{
\item{data}{data frame with date and value}

\item{datecol}{the name (as a character string) of the date column}

\item{valcol}{the name (as a character string) of the PM2.5 value column}
}
\value{
data frame with the daily averages
}
\description{
Calcualte the daily average PM2.5 reading, following CAAQS data completeness rules
}
\examples{
\dontrun{

}
}

