% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_stat.R
\name{pm_daily_avg}
\alias{pm_daily_avg}
\title{Find the daily average for PM2.5}
\usage{
pm_daily_avg(data, dt = "date_time", val = "value", by = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column. Default
\code{"date_time"}}

\item{val}{the name (as a character string) of the PM2.5 value column. Default
\code{"value"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specfify
the id column so that it is retained in the output.}
}
\value{
data frame with the daily averages, can be input into
}
\description{
Find the daily average for PM2.5
}

