% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pm_data_complete.R
\name{pm_data_complete}
\alias{pm_data_complete}
\title{Calculates the percentage of days with readings annually, as well as for each
quarter for each year in a dataset.}
\usage{
pm_data_complete(data, dt = "date_time", val = "value", by = NULL,
  daily_valid = 18, year_valid = 75, q_valid = 60)
}
\arguments{
\item{data}{data frame (likely the result of running
\code{\link{pm_daily_avg}})}

\item{dt}{the name of the date/time column (as a character string). Default
\code{"date_time"}}

\item{val}{the name of the column with daily average PM2.5 values. Default
\code{"avg_24hr"}}

\item{by}{character vector of  grouping variables, probably an id if using
multiple sites. Even if not using multiple sites, you shoud specfify the id
column so that it is retained.}

\item{daily_valid}{The minimum number of hourly readings in a day for the
daily average to be considered valid (default \code{18}).}

\item{year_valid}{The percentage of valid days required in a year (default
75).}

\item{q_valid}{The percentage of valid days required in each quarter
(default 60).}
}
\value{
A data frame with percentage of days with readings annually, as well
 as one for each quarter. Also includes whether or not the annual and
 quarterly requirements are met
}
\description{
Designed to be used with the output from \code{\link{pm_daily_avg}}
}
\seealso{
\code{\link{pm_daily_avg}}
}

