% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_stat.R
\name{pollutant_daily_stat}
\alias{pollutant_daily_stat}
\title{Find the daily statistic of a value by some factor}
\usage{
pollutant_daily_stat(data, dt, val, by = NULL, pollutant_standard,
  stat = max_na, exclude_df = NULL, exclude_df_dt = NULL, digits = 1,
  n_readings_min = 18)
}
\arguments{
\item{data}{Dataframe}

\item{dt}{character the column containing dates}

\item{val}{character the value column}

\item{by}{character the columns to group by}

\item{pollutant_standard}{If this value is exceeded, then exceed column will be true.}

\item{stat}{function to summarise by.}

\item{exclude_df}{a data.frame specifying which date ranges and sites to exclude. 
The data.frame must have to one date column if specific dates or date-times are to be excluded,
and exactly two date or date-time columns if a date range is needed.
must have all the columns of by provided,}

\item{exclude_df_dt}{specifies the one (or two) date columns to be use in the exclude_df data.frame.}

\item{digits}{The number of digits that the value should be rounded to.}

\item{n_readings_min}{The minimum number of readings in a day for the summary statistic to be valid.}
}
\value{
dataframe with summarised statistic
}
\description{
Given a dataframe with one value column, find the maximum value by some date.
}
