% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_98_percentile.R
\name{quantile2}
\alias{quantile2}
\title{Return the rank that should be used to determine the 98th percentile given a 
number of valid days}
\usage{
quantile2(x, probs = 0.98, na.rm = FALSE, names = FALSE, type = "caaqs")
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted.}

\item{probs}{numeric vector of probablities with values in \eqn{[0,1]}. 
Default \code{0.98}}

\item{na.rm}{logical; if true, any \code{NA} and \code{NaN}'s are removed 
from \code{x} before the quantiles are computed. Default \code{FALSE}}

\item{names}{logical; if true, the result has a names attribute. Set to FALSE
for speedup with many probs. Default \code{FALSE}}

\item{type}{\code{"caaqs"} (default) or an integer between 1 and 9 selecting
one of the nine base quantile algorithms be used. See
\code{\link[stats]{quantile}} for details}
}
\value{
A vector of \code{length(probs)}; if \code{names = TRUE}, it has a
  \code{names} attribute
}
\description{
Wraps \code{\link[stats]{quantile}} but with different defaults, adds another
\code{type}, "caaqs", where the percentile (default 0.98) is calculated 
according to the caaqs methods.
}
\seealso{
\code{\link[stats]{quantile}}
}

