% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{quantile2_na}
\alias{quantile2_na}
\title{Find the quantile, return NA if no non-NA values.}
\usage{
quantile2_na(x, probs = 0.98, na.rm = TRUE, names = FALSE,
  type = "caaqs")
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted.}

\item{probs}{numeric vector of probablities with values in \eqn{[0,1]}. 
Default \code{0.98}}

\item{na.rm}{logical; if true, any \code{NA} and \code{NaN}'s are removed 
from \code{x} before the quantiles are computed. Default \code{FALSE}}

\item{names}{logical; if true, the result has a names attribute. Set to FALSE
for speedup with many probs. Default \code{FALSE}}

\item{type}{\code{"caaqs"} (default) or an integer between 1 and 9 selecting
one of the nine base quantile algorithms be used. See
\code{\link[stats]{quantile}} for details}
}
\value{
A vector of \code{length(probs)}; if \code{names = TRUE}, it has a
  \code{names} attribute
}
\description{
Wraps \code{\link{quantile2}} but returns NA if no non-NA values
}
\seealso{
\code{\link{quantile2}}
}

