% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rcaaqs-deprecated}
\alias{rcaaqs-deprecated}
\title{Defunct functions in rcaaqs}
\description{
These functions have been removed from rcaaqs.
}
\details{
\itemize{
 \item \code{\link{format_date}}: This function is defunct. Use \code{\link{format_caaqs_dt}} instead.
}
}
