% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rd_map}
\alias{rd_map}
\title{Borders of British Columbia Regional Districts}
\format{A SpatialPolgonsDataFrame with 29 polygons and 2 variables:
\describe{
\item{unit_name}{The name of the region}
\item{region_type}{The type of region (Regional District, Regional Municipality, or Unincorporated Area)}
}}
\usage{
rd_map
}
\description{
Borders of British Columbia Regional Districts
}
\keyword{datasets}

