% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Compute the rolling mean with different windows and thresholds.}
\usage{
rolling_mean(x, width = 8, thresh = 0)
}
\arguments{
\item{x}{the input numeric value}

\item{width}{The maximum number of values used to compute the mean.}

\item{thresh}{If at least thresh values are not used in computing 
the mean, then the output will be set to NA.}
}
\value{
A numeric vector with the rolling mean. The output vector will
always have the same length as the input vector, which makes it convenient 
to use with the summarise function in dplyr.
}
\description{
Compute the rolling mean with different windows and thresholds.
}

