% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_stat.R
\name{rolling_stat_page}
\alias{rolling_stat_page}
\alias{o3_rolling_8hr_avg}
\alias{so2_three_yr_avg}
\alias{no2_three_yr_avg}
\alias{pm_three_yr_avg}
\alias{o3_three_yr_avg}
\title{Compute a rolling statistic (typically over years, but also hourly).}
\usage{
o3_rolling_8hr_avg(data, dt = "date_time", val = "value", by = NULL)

so2_three_yr_avg(data, dt = "year", val = "ann_99_percentile", by = NULL)

no2_three_yr_avg(data, dt = "year", val = "ann_98_percentile", by = NULL)

pm_three_yr_avg(data, dt = "year", val = "ann_98_percentile", by = NULL)

o3_three_yr_avg(data, dt = "year", val = "max8hr", by = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column.}

\item{val}{the name (as a character string) of the value column.}

\item{by}{character vector of grouping variables in data, probably an id if 
using multiple sites. Even if not using multiple sites, you shoud specify 
the id column so that it is retained in the output.}

\item{exclude_df}{he data.frame that has all the columns in the by
parameter, in addition exactly one or two date columns.}

\item{exclude_df_dt}{a character vector with exactly one or two date
columns.}
}
\value{
dataframe with rolling 8-hour averages.
}
\description{
Compute a rolling statistic (typically over years, but also hourly).
}
