% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rolling_sum}
\alias{rolling_sum}
\title{Compute the rolling sum with different windows.}
\usage{
rolling_sum(x, width)
}
\arguments{
\item{x}{the input numeric value}

\item{width}{The maximum number of values used to compute the sum.}
}
\value{
vector with all input dates and values, interpolated with NAs.
}
\description{
Compute the rolling sum with different windows.
}

