% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{round_caaqs}
\alias{round_caaqs}
\title{Round up from 5 (drop-in replacement for base::round)}
\usage{
round_caaqs(x, digits = 0)
}
\arguments{
\item{x}{vector of numbers to round}

\item{digits}{integer indicating the number of decimal places to round to (default 0)}
}
\value{
numeric vector
}
\description{
CAAQS methods requires always rounding up numbers ending in 5, 
which is different than the default round, which uses the 'round to even'
rule. See ?round
}
