% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.R
\name{set_opts}
\alias{set_opts}
\title{Set options}
\usage{
set_opts(...)
}
\arguments{
\item{...}{options to set. In the form option_name = "option value". Valid
option names are: \code{tz}.}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Set options
}
\details{
Note that this method of setting options does no checking for valid
  option values. For setting the timezone (\code{tz}), it is better to use
  the shortcut function \link{set_tz}.
}
\examples{
set_opts(tz = "Etc/GMT+8")
}

