% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.R
\name{set_tz}
\alias{set_tz}
\title{Set global timezone}
\usage{
set_tz(timezone)
}
\arguments{
\item{timezone}{A valid timezone to use in all date_time operations}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Set global timezone
}
\examples{
set_tz("Etc/GMT+8")
}

