% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{test_time_interval}
\alias{test_time_interval}
\title{Find time interval of a vector of regularly spaced dates/datetimes}
\usage{
test_time_interval(x)
}
\arguments{
\item{x}{vector of dates/datetimes}
}
\value{
character
}
\description{
Find time interval of a vector of regularly spaced dates/datetimes
}
