% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{valid_by_quarter}
\alias{valid_by_quarter}
\title{Calcuate number (or proportion) of valid days in each year and quarter}
\usage{
valid_by_quarter(data, date, by, units = c("prop", "days"))
}
\arguments{
\item{data}{the input data.frame}

\item{date}{the date}

\item{by}{the grouping variable}

\item{units}{either "prop" if you want proportion of valid days or "days" is absolute number of days is required.}
}
\value{
data.frame with valid days in each year and quarter
}
\description{
Calcuate number (or proportion) of valid days in each year and quarter
}

