% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_stat.R
\name{yearly_stat_page}
\alias{no2_avg_hourly_by_year}
\alias{no2_yearly_98}
\alias{o3_ann_4th_highest}
\alias{pm_yearly_98}
\alias{pm_yearly_avg}
\alias{so2_avg_hourly_by_year}
\alias{so2_yearly_99}
\alias{yearly_stat_page}
\title{Compute a yearly statistic (typically 98th percentile).}
\usage{
pm_yearly_98(data, dt = "date", val = "avg_24h", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

so2_yearly_99(data, dt = "date_time", val = "max_24h", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

no2_yearly_98(data, dt = "date_time", val = "max_24h", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

pm_yearly_avg(data, dt = "date", val = "avg_24h", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

o3_ann_4th_highest(data, dt = "date", val = "max8hr", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

so2_avg_hourly_by_year(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)

no2_avg_hourly_by_year(data, dt = "date_time", val = "value", by = NULL,
  exclude_df = NULL, exclude_df_dt = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column. Default
\code{"date"}}

\item{val}{the name (as a character string) of the value column. Default
\code{"value"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specify
the id column so that it is retained in the output.}

\item{exclude_df}{he data.frame that has all the columns in the 
by parameter, in addition exactly one or two date columns.}

\item{exclude_df_dt}{a character vector with exactly one or two date columns.}
}
\value{
data frame with the yearly summary (typically the 98th percentile)
}
\description{
Compute a yearly statistic (typically 98th percentile).
}

