% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airzone_map}
\alias{airzone_map}
\title{Borders of British Columbia Airzones}
\format{A SpatialPolgonsDataFrame with 7 polygons and 1 variable
\describe{
\item{Airzone}{The name of the airzone}
}}
\source{
\url{http://catalogue.data.gov.bc.ca/dataset/british-columbia-air-zones}
}
\usage{
airzone_map
}
\description{
Borders of British Columbia Airzones
}
\keyword{datasets}

