% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/airzone_metric.R
\name{airzone_metric}
\alias{airzone_metric}
\title{Calculate the airzone metric}
\usage{
airzone_metric(df, n_years = "n_years", az = "Airzone", val, keep = NULL)
}
\arguments{
\item{df}{The dataframe}

\item{n_years}{The column containing the number of years each 3yr avg is based
on}

\item{az}{Airzone column}

\item{val}{the column containing the the metric for individual stations}

\item{keep}{columns in the input df that you would like to retain in the
output data frame. You can make it a named vector to rename the column in
the output. Use the form \code{keep = c(new_name = "existing_name")}. This
can also be used to rename any of the columns specified by n_years, az, or
val.}
}
\value{
A dataframe with two columns: metric value and achievement status
}
\description{
<full description>
}

