% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airzone_metric.R
\name{airzone_metric}
\alias{airzone_metric}
\title{Calculate the airzone metric}
\usage{
airzone_metric(
  data,
  n_years = "n_years",
  az = "airzone",
  station_id = "ems_id",
  ambient_metric_val = "metric_value_ambient",
  ambient_caaqs = "caaqs_ambient",
  excluded = "excluded",
  mgmt_metric_val = "metric_value_mgmt",
  mgmt = "mgmt_level",
  keep = NULL
)
}
\arguments{
\item{data}{The dataframe}

\item{n_years}{The column containing the number of years each 3yr avg is
based on}

\item{az}{The airzone column}

\item{station_id}{The column containing the unique station id}

\item{ambient_metric_val}{The column containing the ambient CAAQS metric for individual stations}

\item{ambient_caaqs}{The column containing the CAAQS achievement levels for individual
stations}

\item{excluded}{Logical vector indicating if ambient data was excluded for calculating
the CAAQS management levels}

\item{mgmt_metric_val}{The column containing the management CAAQS metric for individual stations}

\item{mgmt}{The column containing the CAAQS management levels for individual
stations}

\item{keep}{Character vector specifying the columns in the input data that you would like to retain in the
output data frame.}
}
\value{
A dataframe with 9 columns: airzone, number of years in calculation,
ambient CAAQS metric value, CAAQS achievement levels, ambient reporting station id,
logical vector indicating if ambient data was excluded for management calculations,
management CAAQS metric value, CAAQS management levels, and management reporting station id.
}
\description{
Calculate the metrics for each airzone based on station-level results.
Determines the CAAQS achievement levels and CAAQS management levels for each
airzone and reports on the representative stations for these metrics in each
airzone.
}
