% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airzone_metric.R
\name{assign_airzone}
\alias{assign_airzone}
\title{Assign locations to airzones}
\usage{
assign_airzone(data, airzones, az = "Airzone", station_id = "ems_id",
  coords = c("lat", "lon"))
}
\arguments{
\item{data}{Data frame. Contains station ids and air quality data}

\item{airzones}{SpatialPolygonsDataFrame. Polygons reflecting airzone
locations}

\item{az}{Character. Name of airzones column in the 'airzones'
SpatialPolygonsDataFrame object}

\item{station_id}{Character. Name of the station_id column in 'data'}

\item{coords}{Character vector. Names of the columns containing latitude and
longitude (respectively) for each station. Defaults to 'lat' and 'lon'}
}
\description{
With a data set containing station locations, calculate which airzones each
station belongs to. Requires a Spatial Polygons Data Frame object from the
\code{sp} package containing the airzone names and locations (polygons).
}
\details{
For British Columbia, consider using the \code{airzones("sp")} function to
extract a Spatial Polygons Data Frame of BC airzones from the
\code{\link[bcmaps]{bcmaps}} package.
}
\examples{

\dontrun{
# Using the airzone spatial dataframe from the bcmaps package
bc_airzones <- bcmaps::airzones("sp")

by_airzones <- assign_airzone(data, bc_airzones)
}

}
