% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airzone_metric.R
\name{assign_airzone}
\alias{assign_airzone}
\title{Assign locations to airzones}
\usage{
assign_airzone(df, airzones, az = "Airzone", station_id = "ems_id",
  coords = c("lat", "lon"))
}
\arguments{
\item{df}{Data frame. Contains station ids and air quality data.}

\item{airzones}{SpatialPolygonsDataFrame. Polygons reflecting airzone
locations}

\item{az}{Character. Name of airzones column in the 'airzones'
SpatialPolygonsDataFrame object}

\item{station_id}{Character. Name of the station_id column in the 'df' data
frame}

\item{coords}{Character vector. Names of the columns containing latitude and 
longitude (respectively) for each station. Defaults to 'lat' and 'lon'}
}
\description{
With a data set containing station locations, calculate which airzones each 
station belongs to. Requires a Spatial Polygons Data Frame object from the
\code{sp} package containting the airzone names and locations (polygons).
}
\examples{

\dontrun{
# Using the airzone spatial dataframe from the bcmaps package
bc_airzones <- bcmaps::airzones("sp")

df_airzones <- assign_airzone(df, bc_airzones)
}

}
