% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cut_caaq.R
\name{cut_caaq}
\alias{cut_caaq}
\title{Bin a vector of air quaility readings into CAAQS management levels}
\usage{
cut_caaq(x, metric, output = "labels", drop_na = FALSE)
}
\arguments{
\item{x}{vector of air quality readings (pollutant concentrations)}

\item{metric}{What is the metric? Must be one of: \code{"o3", "pm2.5_annual", "pm2.5_daily"}}

\item{output}{should the function output labels (\code{"labels"}; default
or break values in unicode (\code{"breaks_u"}) or break values in html (\code{"breaks_h"})?}

\item{drop_na}{Should NA values be dropped, or retained as factor level "Insufficient Data" (default)?}
}
\value{
factor
}
\description{
Can be used for Ozone, PM2.5 (daily or annual)
}
\examples{
\dontrun{

}
}

