% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pm_98_percentile.R
\name{cut_rank}
\alias{cut_rank}
\title{Return the rank that should be used to determine the 98th percentile given a
number of valid days}
\usage{
cut_rank(n)
}
\arguments{
\item{n}{the number of valid days in a year}
}
\value{
a number from 1 to 8
}
\description{
Return the rank that should be used to determine the 98th percentile given a
number of valid days
}
\examples{
\dontrun{

}
}
\keyword{internal}

