% Generated by roxygen2 (4.0.2): do not edit by hand
\name{date_fill}
\alias{date_fill}
\title{Fill gaps in a date sequence}
\usage{
date_fill(df, date_col, interval = NULL, fill_cols = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{date_col}{the column containing dates}

\item{interval}{The interval in the date sequence. If \code{NULL}, calculated automatically.}

\item{fill_cols}{Columns to fill with the value in the column (should be columns where value is same in every row, such as an ID.)}
}
\value{
dataframe with filled in dates
}
\description{
Given a dataframe with one column as a date sequence, fill gaps in the dat sequence.
}
\examples{
\dontrun{

}
}

