% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dates.R
\name{date_fill}
\alias{date_fill}
\title{Fill gaps in a date sequence}
\usage{
date_fill(df, date_col, interval = NULL, fill_cols = NULL,
  add_ymd = FALSE)
}
\arguments{
\item{df}{Dataframe}

\item{date_col}{the column containing dates}

\item{interval}{The interval in the date sequence. If \code{NULL}, calculated automatically.}

\item{fill_cols}{Columns to fill with the value in the column (should be columns where value is same in every row, such as an ID.)}

\item{add_ymd}{logical. Should the date be split into year, month, and day columns and added to the output?}
}
\value{
dataframe with filled in dates
}
\description{
Given a dataframe with one column as a date sequence, fill gaps in the dat sequence.
}
\examples{
\dontrun{
foo <- data.frame(Date = seq(as.Date("2008-01-01"), as.Date("2008-12-01"), by = "month"),
                  val = round(rnorm(12, 5, 2), 1), label = rep("a", 12))
bar <- foo[-c(2,5,6,7,10),]
date_fill(bar, "Date", interval = "1 month", fill_cols = "label")
}
}

