% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dates.R
\name{find_time_int}
\alias{find_time_int}
\title{Find time interval in a date sequence}
\usage{
find_time_int(dates)
}
\arguments{
\item{dates}{vector of dates}
}
\value{
an integer reflecting the number of seconds in the time interval
}
\description{
Taken from package openair
}
\examples{
\dontrun{

}
}

