% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{format_date}
\alias{format_date}
\title{Convert character date string from Raw air data to POSIXct}
\usage{
format_date(dates, format = "\%Y-\%m-\%d \%H:\%M:\%S", lt = FALSE)
}
\arguments{
\item{dates}{vector of dates to convert as characters.}

\item{format}{the format of the character dates}

\item{lt}{return date object as POSIXlt. Default \code{FALSE}}
}
\value{
dataframe with filled in dates
}
\description{
Likely data downloaded from DataBC
}
