% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{format_date}
\alias{format_date}
\title{This function is defunct. Use \code{\link[=format_caaqs_dt]{format_caaqs_dt()}}}
\usage{
format_date(dates, format, lt)
}
\arguments{
\item{dates}{defunct function. do not use}

\item{format}{defunct function. do not use}

\item{lt}{defunct function. do not use}
}
\description{
This function is defunct. Use \code{\link[=format_caaqs_dt]{format_caaqs_dt()}}
}
\seealso{
\code{\link[=format_caaqs_dt]{format_caaqs_dt()}}
}
