% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dates.R
\name{format_date}
\alias{format_date}
\title{Convert character date string from ENVISTA to POSIXct}
\usage{
format_date(dates)
}
\arguments{
\item{dates}{vector of dates to convert. Dates must be in the format "ddMonYYYY:HH:MM:SS"}
}
\value{
dataframe with filled in dates
}
\description{
Convert character date string from ENVISTA to POSIXct
}

