% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut.R
\name{get_levels}
\alias{get_levels}
\title{Return CAAQS levels for any or all, parameters, for either achievement or
management reporting}
\usage{
get_levels(type, parameter = "all")
}
\arguments{
\item{type}{one of "achievement", "management"}

\item{parameter}{one of "all" (default), "o3", "PM2.5_annual", "PM2.5_24h"}
}
\value{
A data frame
}
\description{
Return CAAQS levels for any or all, parameters, for either achievement or
management reporting
}
\examples{
\dontrun{
get_levels("o3")
}
}

