% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{n_within_window}
\alias{n_within_window}
\title{number within a rolling window and step size (for dates or numeric)}
\usage{
n_within_window(x, interval, window)
}
\arguments{
\item{x}{the date, date_time, or numeric value}

\item{interval}{some numeric step size. for date-times, in seconds.}

\item{window}{some numeric look-back window in \code{interval} units.}
}
\value{
vector of same length as x, that returns, for each x, the number 
behind x that were within \code{window*interval} units, including x itself.
}
\description{
number within a rolling window and step size (for dates or numeric)
}
\examples{
rcaaqs:::n_within_window(c(1,2,7,8), interval = 1, window = 2)
# c(1,2,1,2)
}

