% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_stat.R
\name{o3_rolling_8hr_avg}
\alias{o3_rolling_8hr_avg}
\title{Find the daily maximum of a value by some factor}
\usage{
o3_rolling_8hr_avg(data, dt = "date_time", val = "value", by = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{dt}{character the column containing date-times}

\item{val}{character the value column}

\item{by}{character the columns to group by}
}
\description{
Given a dataframe with one value column, find the maximum value by some date.
}

