% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot daily time series data with exceedances}
\usage{
plot_ts(daily_df, parameter, rep_yr, plot_exceedances = FALSE)
}
\arguments{
\item{daily_df}{a dataframe of daily aggregated air quality readings with
columns: date, avg_24h (if pm25), max8hr (if o3)}

\item{parameter}{air pollutant ("o3", "pm2.5_annual", "pm2.5_daily")}

\item{plot_exceedances}{logical. Should exceedances be plotted?}

\item{rep_year}{the reporting year}
}
\value{
a ggplot2 object
}
\description{
Plot daily time series data with exceedances
}

