% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot daily time series data with exceedances and optionally caaqs for a
particular station}
\usage{
plot_ts(x, id = NULL, id_col = NULL, rep_yr, plot_caaqs = TRUE,
  plot_exceedances = FALSE, base_size = 10, annot_size = NULL)
}
\arguments{
\item{x}{an object of class \code{caaqs}}

\item{id}{the id of the station you would like to plot}

\item{id_col}{the column in which to look for \code{id}}

\item{rep_yr}{The reporting year}

\item{plot_caaqs}{Should the caaqs be plotted?}

\item{plot_exceedances}{logical. Should exceedances be plotted?}

\item{base_size}{base font size for the plot}

\item{annot_size}{size of annotations. Scaling of this size is not the same
as the rest of the font sizes, so you will have to experiment. Defaults to
0.32*base_size}
}
\value{
a ggplot2 object
}
\description{
Plot daily time series data with exceedances and optionally caaqs for a
particular station
}
