% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot daily time series data with exceedances}
\usage{
plot_ts(daily_data, caaqs_data = NULL, parameter, rep_yr,
  plot_exceedances = FALSE, base_size = 10, annot_size = NULL)
}
\arguments{
\item{daily_data}{a dataframe of daily aggregated air quality readings with 
columns: date, avg_24h (if pm25), max8hr (if o3)}

\item{caaqs_data}{(optional) a one-row dataframe with columns "min_year", 
"max_year" and columns for caaqs achievement status and mangaement status}

\item{parameter}{air pollutant ("o3", "pm2.5_annual", "pm2.5_24h")}

\item{rep_yr}{The reporting year}

\item{plot_exceedances}{logical. Should exceedances be plotted?}

\item{base_size}{base font size for the plot}

\item{annot_size}{size of annotations. Scaling of this size is not the same
as the rest of the font sizes, so you will have to experiment. Defaults to
0.32*base_size}
}
\value{
a ggplot2 object
}
\description{
Plot daily time series data with exceedances
}
