% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caaqs.R
\name{pm_24h_caaq}
\alias{pm_24h_caaq}
\title{Calculate the PM2.5 24 hour CAAQ metric}
\usage{
pm_24h_caaq(data, year = "year", val = "pm_metric", by = NULL)
}
\arguments{
\item{data}{Data frame. Contains three year PM2.5 24hr average data output from
\code{pm_three_yr_avg}}

\item{year}{Character. The name of the data column containing years. Defaults
to 'year'.}

\item{val}{Character. The name of the data column containing the pm metric.
Defaults to 'pm_metric'.}

\item{by}{Character. The name(s) of the data column(s) specifying the data
grouping variables (i.e. site_id, etc.). Even if not using multiple sites,
you shoud specfify the id column so that it is retained in the output.}
}
\value{
A data frame arranged by grouping variables (if present) with caaq
  metrics, achievement levels and management levels.
}
\description{
Calculates and returns the PM2.5 24 hour CAAQ metric based on a rolling
three-year average.
}
