% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pm_24hr_caaq.R
\name{pm_24hr_caaq}
\alias{pm_24hr_caaq}
\title{Calculates the three year average for the PM2.5 24 hour metric}
\usage{
pm_24hr_caaq(data, yearcol, valcol, ..., year = "latest")
}
\arguments{
\item{data}{data frame containing annual 98th percentile values, most likely
generated by \code{\link{pm_98_percentile}}}

\item{yearcol}{name of the column containing the year}

\item{valcol}{name of the column containing annual 98th percentile values}

\item{...}{grouping variables in data, probably an id if using multiple
sites. Even if not using multiple sites, you shoud specfify the id column
so that it is retained in the output.}

\item{year}{The year to calculate the metric for (this will be the latest of
the three years used in the calculation). Can be "latest" to use the latest
year found in the data frame, or an integer denoting the year.}
}
\value{
a data frame, with year, the value, and the number of years the value
  is based on
}
\description{
Flags if based on only two years.
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{pm_98_percentile}}
}

