% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pm_98_percentile}
\alias{pm_98_percentile}
\title{Calculate the annual 98th percentile of daily average PM2.5 values according to CAAQS standards}
\usage{
pm_98_percentile(data, datecol, valcol, ...)
}
\arguments{
\item{data}{data frame}

\item{datecol}{the name of the "year" column (as a character string)}

\item{valcol}{the name of the column with daily average PM2.5 values}

\item{...}{grouping variables, probably an id if using multiple sites}
}
\value{
A data frame with 98th percentiles of daily averages, per year
}
\description{
Designed to be used with the output from \code{\link{pm_avg_daily}}
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{pm_avg_daily}}
}

