% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pm_98_percentile.R
\name{pm_98_percentile}
\alias{pm_98_percentile}
\title{Calculate the annual 98th percentile of daily average PM2.5 values according
to CAAQS standards}
\usage{
pm_98_percentile(data, datecol, valcol, by = NULL, std = 28,
  completeness = TRUE, year_valid = 75, q_valid = 60)
}
\arguments{
\item{data}{data frame}

\item{datecol}{the name of the "date" column (as a character string)}

\item{valcol}{the name of the column with daily average PM2.5 values}

\item{by}{character vector of grouping variables in data, probably an id if using multiple
sites. Even if not using multiple sites, you shoud specfify the id column
so that it is retained in the output.}

\item{std}{the value of the PM2.5 standard (default 28).}

\item{completeness}{Should the completeness criteria be calculated (default
TRUE)}

\item{year_valid}{The percentage of valid days required in a year (default
75). Only required if calculating the completeness criteria.}

\item{q_valid}{The percentage of valid days required in each quarter
(default 60). Only required if calculating the completeness criteria.}
}
\value{
A data frame with 98th percentiles of daily averages, per year
}
\description{
Also computes the data completeness criteria (optional). Designed to be used
with the output from \code{\link{pm_avg_daily}}.
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{pm_avg_daily}}
}

