% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_annual_caaq.R
\name{pm_annual_caaq}
\alias{pm_annual_caaq}
\title{Calculate the annual CAAQS metric for PM2.5}
\usage{
pm_annual_caaq(data, year = "year", val = "ann_avg", by = NULL,
  cyear = "latest")
}
\arguments{
\item{data}{data frame containing annual average values}

\item{year}{name of the column containing the year. Default \code{"year"}}

\item{val}{name of the column containing annual average values. Default
\code{"ann_avg"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specfify
the id column so that it is retained in the output.}

\item{cyear}{The year to calculate the metric for (this will be the latest of
the three years used in the calculation). Can be "latest" to use the latest 
year found in the data frame, or an integer denoting the year.}
}
\value{
A data frame
}
\description{
Calculate the annual CAAQS metric for PM2.5
}
