% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pm_avg_daily.R
\name{pm_avg_daily}
\alias{pm_avg_daily}
\title{Calcualte the daily average PM2.5 reading, following CAAQS data completeness rules}
\usage{
pm_avg_daily(data, datecol, valcol, ...)
}
\arguments{
\item{data}{data frame with date and value}

\item{datecol}{the name (as a character string) of the date column}

\item{valcol}{the name (as a character string) of the PM2.5 value column}

\item{...}{grouping variables, probably an id if using multiple sites}
}
\value{
data frame with the daily averages, can be input into \code{\link{pm_98_percentile}}
}
\description{
Calcualte the daily average PM2.5 reading, following CAAQS data completeness rules
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{pm_98_percentile}}
}

