% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pm_data_complete.R
\name{pm_data_complete}
\alias{pm_data_complete}
\title{Calculates the percentage of days with readings annually, as well as for each
quarter for each year in a dataset.}
\usage{
pm_data_complete(data, datecol, valcol, ..., year_valid = 75, q_valid = 60)
}
\arguments{
\item{data}{data frame (likely the result of running \code{\link{pm_avg_daily}})}

\item{datecol}{the name of the "date" column (as a character string)}

\item{valcol}{the name of the column with daily average PM2.5 values}

\item{...}{grouping variables, probably an id if using multiple sites. Even
if not using multiple sites, you shoud specfify the id column so
that it is retained.}

\item{year_valid}{The percentage of valid days required in a year (default 75).
Must be named as it comes after grouping variables (...)}

\item{q_valid}{The percentage of valid days required in each quarter (default 60).
Must be named as it comes after grouping variables (...)}
}
\value{
A data frame with percentage of days with readings annually, as well
        as one for each quarter. Also includes whether or not the annual and
        quarterly requirements are met
}
\description{
Designed to be used with the output from \code{\link{pm_avg_daily}}
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{pm_avg_daily}}
}

