% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_stat.R
\name{pm_yearly_98}
\alias{pm_yearly_98}
\title{Find the yearly 98th percentile of the daily average for PM2.5}
\usage{
pm_yearly_98(data, dt = "date", val = "avg_24h", by = NULL)
}
\arguments{
\item{data}{data frame with date and value}

\item{dt}{the name (as a character string) of the date-time column. Default
\code{"date"}}

\item{val}{the name (as a character string) of the value column. Default
\code{"value"}}

\item{by}{character vector of grouping variables in data, probably an id if
using multiple sites. Even if not using multiple sites, you shoud specify
the id column so that it is retained in the output.}
}
\value{
data frame with the 98th percentile
}
\description{
Find the yearly 98th percentile of the daily average for PM2.5
}

