% Generated by roxygen2 (4.0.2): do not edit by hand
\name{poly_summary}
\alias{poly_summary}
\title{A function to count the number of unique occurrences of a variable within a polygon}
\usage{
poly_summary(polys, points, fn, ...)
}
\arguments{
\item{polys}{SpatialPolygonsDataFrame to summarise by}

\item{points}{SpatialPointsDataFrame to summarise}

\item{fn}{The summary function with which to summarise point data in each polygon}

\item{var}{Variable (column name) in \code{points} to summarise}
}
\value{
A vector of the same length as the SpatialPolygonsDataFrame.
}
\description{
<full description of function>
}
\examples{
\dontrun{

}
}

