% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_summary.R
\name{poly_summary}
\alias{poly_summary}
\title{A function to summarise occurrences of a variable within a polygon}
\usage{
poly_summary(polys, points, fn, ...)
}
\arguments{
\item{polys}{SpatialPolygonsDataFrame to summarise by}

\item{points}{SpatialPointsDataFrame to summarise}

\item{fn}{The summary function with which to summarise point data in each polygon}

\item{...}{Other arguments to fn}
}
\value{
A data frame with the same number of rows as the SpatialPolygonsDataFrame.
}
\description{
A function to summarise occurrences of a variable within a polygon
}
