% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcaaqs.R
\docType{package}
\name{rcaaqs}
\alias{rcaaqs}
\title{rcaaqs: Tools for calculating CAAQS with air quality data}
\description{
The rcaaqs package provides functions for the calculation of Canadian Ambient
Air Quality Standards (CAAQS).
}
\section{rcaaqs functions}{


\itemize{
\item General functions for doing things like formatting dates,
filling in sequences, etc.
\item Functions for step-wise calculation of CAAQS metrics for different
pollutants. Currently these are complete for fine particulate matter (PM2.5),
ground-level ozone (O3), sulphur dioxide (SO2) and nitrogen dioxide (NO2).
\item Functions for assigning metrics for different pollutants into
achievement and management categories.
}
}

