% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rd_map}
\alias{rd_map}
\title{Borders of British Columbia Regional Districts}
\format{A SpatialPolgonsDataFrame with 29 polygons and 2 variables:
\describe{
\item{region_name}{The name of the region}
\item{region_type}{The type of region (Regional District, Regional Municipality, or Unincorporated Area)}
}}
\source{
\url{http://catalogue.data.gov.bc.ca/dataset/ceei-primary-indicators-total-2007-regional-districts}
}
\usage{
rd_map
}
\description{
Borders were taken from the Community Energy Emissions Inventory shapefile
available from the source below. Northern Rockies Regional Municipality was
added, as it was not available in the original source file.
}
\keyword{datasets}

