% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Compute the rolling mean with different windows and thresholds.}
\usage{
rolling_mean(x, width = 8, thresh = 0)
}
\arguments{
\item{x}{the input numeric value}

\item{width}{The maximum number of values used to compute the mean.}

\item{thresh}{If at least thresh values are not used in computing 
the mean, then the output will be set to NA.}
}
\value{
A rolling mean (see \code{\link[rcaaqs]{rolling_sum}}), except  
if not at least thresh values, it returns NA.
}
\description{
Compute the rolling mean with different windows and thresholds.
}
