% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Generate a summary plot of individual station CAAQS values, grouped by Airzone}
\usage{
summary_plot(data, metric_val, station, airzone, parameter,
  base_size = 12, pt_size = 4, az_labeller = label_value, ...)
}
\arguments{
\item{data}{a data frame with columns for the metric value, the station name,
and the air zone}

\item{metric_val}{name of column in data containing the CAAQS metric value
for each station}

\item{station}{name of column in data containing monitoring station names}

\item{airzone}{name of column in data containing arizones}

\item{parameter}{the name of the column containing the parameter(s) of interest}

\item{base_size}{base font size (default 12)}

\item{pt_size}{size of points (default 4)}

\item{az_labeller}{ggplot2 labeller function to customize the airzone labels.
Default \code{label_value} (i.e., no formatting of Airzone names).
See \code{\link[ggplot2:label_value]{ggplot2::label_value()}}}

\item{...}{options passed on to theme()}
}
\value{
ggplot2 object
}
\description{
Generate a summary plot of individual station CAAQS values, grouped by Airzone
}
