% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Gnerate a summary plot of individual station CAAQS values, grouped by Airzone}
\usage{
summary_plot(data, metric_val, station, airzone, parameter, base_size = 12,
  pt_size = 4, ...)
}
\arguments{
\item{data}{a data frame with columns for the metric value, the station name,
and the air zone}

\item{metric_val}{name of column in data containing the CAAQS metric value
for each station}

\item{station}{name of column in data containing monitoring station names}

\item{airzone}{name of column in data containing arizones}

\item{parameter}{the name of the column containing the parameter(s) of interest}

\item{base_size}{base font size (default 12)}

\item{pt_size}{size of points (default 4)}

\item{...}{options passed on to theme()}
}
\value{
ggplot2 object
}
\description{
Gnerate a summary plot of individual station CAAQS values, grouped by Airzone
}

